feat: Add productPosted and userGrowthPercentage states to AreaCards component
This commit adds the productPosted and userGrowthPercentage states to the AreaCards component. The productPosted state is used to track the number of products posted by users, fetched from the API endpoint "https://baskom-api.up.railway.app/api/v1/products". The userGrowthPercentage state calculates the growth percentage of active users compared to the previous count. These states are used to display the growth of products and users in the AreaCards component.