feat: Add logic to upgrade user role and add new role to user
This commit adds logic to upgrade the user role in the UpgradeRoleModalAcc component. It sends a PUT request to the API endpoint "https://baskom-api.up.railway.app/api/v1/upgrade-roles/{upgradeRoleId}" with the status set to "accept". Additionally, it sends a POST request to the API endpoint "https://baskom-api.up.railway.app/api/v1/user/roles/add" to add a new role (roleId: 3) to the user identified by userId. These changes ensure that the user's role is upgraded and the new role is added successfully.